refactor(board): move board test helpers under features/board/testing
Board-specific fixtures and DB seeding utilities lived scattered across
src/shared/testing and features/board/storage; consolidate them into a
single features/board/testing module alongside a barrel export.
