+++ src/app/routing/loaders/root-index-loader.test.ts
@@ -4,7 +4,8 @@ import { getBoardSets } from "@features/board";
 import { resetBoardsDB, seedBoardSets } from "@features/board/testing";
 import { rootIndexLoader } from "./root-index-loader";
 
-const FIXTURE_BOARD_URL = "/src/shared/testing/sample-boards/lots-of-stuff.obz";
+const FIXTURE_BOARD_URL =
+  "/src/features/board/testing/sample-boards/lots-of-stuff.obz";
 const DEFAULT_BOARD_PATH = "/quick-core-24.obz";
 
 function callLoader(searchParams = ""): Promise<Response> {
