+++ src/features/board/import/use-board-file-drop.test.tsx
@@ -1,10 +1,9 @@
 import { AppProviders } from "@shared/providers/app-providers";
-import { loadFixtureFile } from "@shared/testing/fixtures";
 import { MemoryRouter, useLocation } from "react-router";
 import { beforeEach, describe, expect, test, vi } from "vitest";
 import { render } from "vitest-browser-react";
 import { listBoardSets } from "../storage/boards-db";
-import { resetBoardsDB } from "../storage/test-utils";
+import { loadFixtureFile, resetBoardsDB } from "../testing";
 import { useBoardFileDrop } from "./use-board-file-drop";
 
 const OBF_FIXTURE = "lots-of-stuff.obf";
