+++ src/features/board/import/use-file-handler-launch.test.tsx
@@ -1,10 +1,9 @@
 import { AppProviders } from "@shared/providers/app-providers";
-import { loadFixtureFile } from "@shared/testing/fixtures";
 import { MemoryRouter, useLocation } from "react-router";
 import { beforeEach, describe, expect, test, vi } from "vitest";
 import { render } from "vitest-browser-react";
 import { listBoardSets } from "../storage/boards-db";
-import { resetBoardsDB } from "../storage/test-utils";
+import { loadFixtureFile, resetBoardsDB } from "../testing";
 import { useFileHandlerLaunch } from "./use-file-handler-launch";
 
 const OBF_FIXTURE = "lots-of-stuff.obf";
