+++ src/features/board/tile/tile.test.tsx
@@ -1,6 +1,6 @@
-import { TEST_IMAGE_SRC } from "@shared/testing/fixtures";
 import { describe, expect, test, vi } from "vitest";
 import { render } from "vitest-browser-react";
+import { TEST_IMAGE_SRC } from "../testing";
 import { Tile } from "./tile";
 
 describe("Tile", () => {
